- [add] 修改后台救助发布管理，增加相册，增加救助详情接口

+++ api-admin/src/main/java/com/hailu/cloud/api/admin/module/xinan/service/XaHelpMenberService.java
@@ -20,7 +20,7 @@ public interface XaHelpMenberService {
      * @param xaHelpMember
      * @return
      */
-    int insert(XaHelpMember xaHelpMember);
+    Long insert(XaHelpMember xaHelpMember);
 
     /**
      * 更新方法
